CollectionListView gets specific collection as prop
The issue now is that each time we reload the page again then the singleCollectionArray is emptied. This is fine, as this functionality is only used whenever we enter the collection page from the myCollection page. This saves us one API call, but we need to fix the problem of reloading the page and keeping the data persistent