Implement persistencetoModel for myCollections
We are now able to persist the data in the store. There are other issues with the presenters not using this data obviously. We still need to make the presenters make use of this dynamic data that is being read from firebase, as well as letting the users actually add to a collection.

We also have to implement (MAYBE) persistenceToModelForSingleCollection since we have the functionality of pressing on a single collection and viewing it.
Although on second thought, this might not actually be necessary as we already have all the collections stored in singleCollectionsArray now and we could simply us a useSelector or something to grab the latest datra and filter out the collection that we need or something similar.